requirements.txt
# Core dependencies
ccxt>=2.0.0
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.2.0
tensorflow>=2.10.0
# OR pytorch>=1.13.0 (choose one)

# Data visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Technical analysis
ta>=0.10.0

# Utilities
python-dotenv>=0.21.0
pyyaml>=6.0
requests>=2.28.0

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0

# Logging
loguru>=0.6.0